# swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "MyFlix Express API with Swagger",
    "version": "2.0.0",
    "description": "This is a simple CRUD API application that serves movie data, made with Express and documented with Swagger"
  },
  "servers": [
    {
      "url": "https://high-triode-348322.lm.r.appspot.com/"
    }
  ],
  "paths": {
...
    "/movies/director/{directorName}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "directorName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The name of director."
          }
        ],
        "responses": {
          "200": {
            "description": "Details of the director that was searched by name.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "type": {
                          "type": "string",
                          "example": "object"
                        },
                        "properties": {
                          "type": "object",
                          "properties": {
                            "name": {
                              "type": "object",
                              "properties": {
                                "type": {
                                  "type": "string",
                                  "example": "string"
                                },
                                "description": {
                                  "type": "string",
                                  "example": "The name of the director."
                                }
                              }
                            },
                            "bio": {
                              "type": "object",
                              "properties": {
                                "type": {
                                  "type": "string",
                                  "example": "string"
                                },
                                "description": {
                                  "type": "string",
                                  "example": "A brief life history of the director."
                                }
                              }
                            },
                            "birthyear": {
                              "type": "object",
                              "properties": {
                                "type": {
                                  "type": "string",
                                  "example": "string"
                                },
                                "description": {
                                  "type": "string",
                                  "example": "The birthyear of the director."
                                }
                              }
                            },
                            "deathyear": {
                              "type": "object",
                              "properties": {
                                "type": {
                                  "type": "string",
                                  "example": "string"
                                },
                                "description": {
                                  "type": "string",
                                  "example": "The deathyear of the director."
                                }
                              }
                            }
                          }
                        }
                      }
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "type": {
                          "type": "string",
                          "example": "object"
                        },
                        "properties": {
                          "type": "object",
                          "properties": {
                            "name": {
                              "type": "object",
                              "properties": {
                                "type": {
                                  "type": "string",
                                  "example": "string"
                                },
                                "description": {
                                  "type": "string",
                                  "example": "The name of the director."
                                }
                              }
                            },
                            "bio": {
                              "type": "object",
                              "properties": {
                                "type": {
                                  "type": "string",
                                  "example": "string"
                                },
                                "description": {
                                  "type": "string",
                                  "example": "A brief life history of the director."
                                }
                              }
                            },
                            "birthyear": {
                              "type": "object",
                              "properties": {
                                "type": {
                                  "type": "string",
                                  "example": "string"
                                },
                                "description": {
                                  "type": "string",
                                  "example": "The birthyear of the director."
                                }
                              }
                            },
                            "deathyear": {
                              "type": "object",
                              "properties": {
                                "type": {
                                  "type": "string",
                                  "example": "string"
                                },
                                "description": {
                                  "type": "string",
                                  "example": "The deathyear of the director."
                                }
                              }
                            }
                          }
                        }
                      }
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}